server/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.2.1
beautifulsoup4==4.13.3
blis==0.7.11
breadability==0.1.20
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
google-api-core==2.24.2
google-api-python-client==2.117.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.69.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.25.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
Mako==1.3.9
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.2
preshed==3.0.9
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.13.1
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
srsly==2.5.1
starlette==0.46.1
sumy==0.11.0
thinc==8.2.5
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.9.4
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.27.1
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.4
weasel==0.3.4
websockets==15.0.1
